pyproject.toml
[tool.poetry]
name = "infinity-tools"
version = "0.1.0"
description = ""
authors = ["Daniel Hensley <daniel@toinfinity.ai>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
numpy = "^1.22.3"
pandas = "^1.4.2"
matplotlib = "^3.5.1"
requests = "^2.27.1"
pyserde = "^0.7.2"
plotly = "^5.7.0"
imageio = "^2.18.0"
ipython = "^8.3.0"
tensorflow = "2.7"
tensorflow-hub = "^0.12.0"
opencv-python = "^4.5.5"
pycocotools = "^2.0.4"
kaleido = "0.2.1"
tqdm = "^4.64.0"
jupyter = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
